[GreenD] * Thêm các component và màng mới(Invoice, Contract...)

+++ src/screens/customer/AddNewCustomerScreen.jsx
@@ -1,6 +1,6 @@
 import { View, Text, Image, Alert, FlatList } from 'react-native'
-import React, { useEffect, useState } from 'react'
-import { ButtonComponent, CardComponent, CircleComponent, ContainerComponent, HeaderComponent, ImagePickerComponent, LoadingModalComponent, SectionComponent, SpaceComponent, TextComponent } from '../../components'
+import React, { useEffect, useRef, useState } from 'react'
+import { ButtonComponent, CardComponent, CircleComponent, ContainerComponent, DateTimePickerComponent, HeaderComponent, ImagePickerComponent, LoadingModalComponent, SectionComponent, SpaceComponent, TextComponent } from '../../components'
 import InputComponent from '../../components/InputComponent'
 import { apiCustomer, apiRoom } from '../../apis/apiDTHome'
 import { showMessage } from 'react-native-flash-message'
@@ -8,6 +8,7 @@ import { appColors } from '../../constants/appColors'
 import { Edit, Trash } from 'iconsax-react-native'
 import storage from '@react-native-firebase/storage';
 import { appInfors } from '../../constants/appInfors'
+import { images } from '../../constants/images'
 
 const testData = [
     {
@@ -27,28 +28,48 @@ const initCustomer = {
     "email": "",
     "phoneNumber": "",
     "photoUrl": "",
-    "dateOfBirth": "",
+    "dateOfBirth": new Date(),
     "citizenIdphotoFirstUrl": "",
     "citizenIdphotoBackUrl": "",
     "anotherPhotoUrl": "",
-    "createdAt": "",
-    "updatedAt": ""
+    createdAt: new Date(),
+    updatedAt: new Date(),
 }
 
 const AddNewCustomerScreen = ({ navigation, route }) => {
     const { customerId, actionType } = route.params
     const [dataCustomer, setDataCustomer] = useState(initCustomer)
     const [imageSelected, setImageSelected] = useState(null)
+    const [imageSelected2, setImageSelected2] = useState(null)
+    const [imageSelected3, setImageSelected3] = useState(null)
     const [isLoading, setIsLoading] = useState(false)
+
+    const refImage = useRef()
+
+    const openImagePicker = () => {
+
+    }
 
     useEffect(() => {
         if (actionType === 'update') {
             fetchDataRoom()
         }
     }, [])
+
+    //quản lý thông tin khách hàng thay đổi
     useEffect(() => {
         console.log(dataCustomer)
     }, [dataCustomer])
+    useEffect(() => {
+        console.log('hình avatar dc chọn nè: ', imageSelected)
+    }, [imageSelected])
+    useEffect(() => {
+        console.log('cccd trước: ', imageSelected2)
+    }, [imageSelected2])
+    useEffect(() => {
+        console.log('cccd sau: ', imageSelected3)
+    }, [imageSelected3])
+    //quản lý thông tin khách hàng thay đổi
 
     const handleChangeValue = (key, value) => {
         let tempData = { ...dataCustomer }
@@ -58,7 +79,7 @@ const AddNewCustomerScreen = ({ navigation, route }) => {
 
     const fetchDataRoom = async () => {
         try {
-            const res = await apiRoom(`/${customerId}`)
+            const res = await apiCustomer(`/${customerId}`)
             setDataCustomer(res)
         }
         catch (e) {
@@ -66,10 +87,11 @@ const AddNewCustomerScreen = ({ navigation, route }) => {
         }
     }
 
-    const upLoadImage = async () => {
+    const upLoadImage = async (imageSelected) => {
         try {
+            const randomNum = Math.floor(Math.random() * (999 - 100 + 1)) + 100;
             const fileExtension = imageSelected.mime.split('/')[1]; // lấy phần đuôi tệp từ mime
-            const fileName = `customer-${imageSelected.modificationDate}.${fileExtension}`;
+            const fileName = `${randomNum}${imageSelected.modificationDate}.${fileExtension}`;
             const pathFireBase = `images/customers/${fileName}`;
 
             // Upload file lên Firebase
@@ -102,18 +124,28 @@ const AddNewCustomerScreen = ({ navigation, route }) => {
         const url = '/create'
         setIsLoading(true)
         try {
-            const downloadURL = await upLoadImage();
-            if (downloadURL) {
-                handleChangeValue('photoUrl', downloadURL);
-                await apiCustomer(url, { ...dataCustomer, photoUrl: downloadURL }, 'post')
-                navigation.navigate('CustomerScreen', { customerUpdate: true })
-                showMessage({
-                    message: "Thông báo",
-                    description: "Thêm khách thành công",
-                    type: "success",
-                })
+            let newDataCustomer = { ...dataCustomer }
+            if (imageSelected) {
+                const avatarUrl = await upLoadImage(imageSelected);
+                newDataCustomer = { ...dataCustomer, photoUrl: avatarUrl }
+            }
+            if (imageSelected2) {
+                const cccdFontUrl = await upLoadImage(imageSelected2);
+                newDataCustomer = { ...dataCustomer, citizenIdphotoFirstUrl: cccdFontUrl }
+            }
+            if (imageSelected3) {
+                const cccdBackUrl = await upLoadImage(imageSelected3);
+                newDataCustomer = { ...dataCustomer, citizenIdphotoBackUrl: cccdBackUrl }
             }
 
+            await apiCustomer(url, newDataCustomer, 'post')
+            navigation.navigate('CustomerScreen', { customerUpdate: true })
+            showMessage({
+                message: "Thông báo",
+                description: "Thêm khách thành công",
+                type: "success",
+            })
+            setIsLoading(false)
         }
         catch {
             showMessage({
@@ -121,6 +153,7 @@ const AddNewCustomerScreen = ({ navigation, route }) => {
                 description: "Thêm khách thất bại",
                 type: "danger",
             })
+            setIsLoading(false)
         }
     }
 
@@ -128,18 +161,28 @@ const AddNewCustomerScreen = ({ navigation, route }) => {
         setIsLoading(true)
         try {
 
-            const downloadURL = await upLoadImage();
-            if (downloadURL) {
-                const newDataCustomer = { ...dataCustomer, photoUrl: downloadURL, updatedAt: new Date() }
-                await apiCustomer(`/update/${customerId}`, newDataCustomer, 'put')
-                navigation.navigate('CustomerScreen', { customerUpdate: true })
-                showMessage({
-                    message: "Thông báo",
-                    description: "Sửa thông tin khách thành công",
-                    type: "success",
-                })
+            let newDataCustomer = { ...dataCustomer, updatedAt: new Date() }
+            if (imageSelected) {
+                const avatarUrl = await upLoadImage(imageSelected);
+                newDataCustomer = { ...dataCustomer, photoUrl: avatarUrl }
+            }
+            if (imageSelected2) {
+                const cccdFontUrl = await upLoadImage(imageSelected2);
+                newDataCustomer = { ...dataCustomer, citizenIdphotoFirstUrl: cccdFontUrl }
+            }
+            if (imageSelected3) {
+                const cccdBackUrl = await upLoadImage(imageSelected3);
+                newDataCustomer = { ...dataCustomer, citizenIdphotoBackUrl: cccdBackUrl }
             }
 
+            await apiCustomer(`/update/${customerId}`, newDataCustomer, 'put')
+            navigation.navigate('CustomerScreen', { customerUpdate: true })
+            showMessage({
+                message: "Thông báo",
+                description: "Sửa thông tin khách thành công",
+                type: "success",
+            })
+            setIsLoading(false)
         }
         catch {
             showMessage({
@@ -147,6 +190,7 @@ const AddNewCustomerScreen = ({ navigation, route }) => {
                 description: "Sửa thông tin khách thất bại",
                 type: "danger",
             })
+            setIsLoading(false)
         }
     }
 
@@ -165,15 +209,33 @@ const AddNewCustomerScreen = ({ navigation, route }) => {
                     onPress: async () => {
                         setIsLoading(true);
                         try {
-
+                            //xủ lý xoá ảnh
                             try {
-                                const fileRef = storage().refFromURL(dataRoom.photoUrl);
-                                if (fileRef) {
-                                    await fileRef.delete();
-                                    console.log('Ảnh đã được xoá khỏi Firebase Storage');
+                                if (dataCustomer.photoUrl) {
+                                    const fileRef = storage().refFromURL(dataCustomer.photoUrl);
+                                    if (fileRef) {
+                                        await fileRef.delete();
+                                        console.log('Ảnh đã được xoá khỏi Firebase Storage');
+                                    }
                                 }
 
-                            } catch (error) {
+                                if (dataCustomer.citizenIdphotoFirstUrl) {
+                                    const fileRef = storage().refFromURL(dataCustomer.citizenIdphotoFirstUrl);
+                                    if (fileRef) {
+                                        await fileRef.delete();
+                                        console.log('Ảnh đã được xoá khỏi Firebase Storage');
+                                    }
+                                }
+
+                                if (dataCustomer.citizenIdphotoBackUrl) {
+                                    const fileRef = storage().refFromURL(dataCustomer.citizenIdphotoBackUrl);
+                                    if (fileRef) {
+                                        await fileRef.delete();
+                                        console.log('Ảnh đã được xoá khỏi Firebase Storage');
+                                    }
+                                }
+                            }
+                            catch (error) {
                                 console.error('Lỗi khi xoá ảnh: ', error);
                             }
 
@@ -184,6 +246,7 @@ const AddNewCustomerScreen = ({ navigation, route }) => {
                                 description: "Xoá khách hàng thành công",
                                 type: "success",
                             });
+                            setIsLoading(false)
                         } catch (e) {
                             console.log('Xoá khách hàng thất bại');
                             showMessage({
@@ -191,6 +254,7 @@ const AddNewCustomerScreen = ({ navigation, route }) => {
                                 description: "Xoá khách hàng thất bại",
                                 type: "danger",
                             });
+                            setIsLoading(false)
                         }
                         setIsLoading(false);
                     },
@@ -203,7 +267,7 @@ const AddNewCustomerScreen = ({ navigation, route }) => {
 
     const renderTestData = ({ item, index }) => {
         return (
-            <CardComponent style={{ alignItems: 'center' }} onPress={()=>{Alert.alert('chỉnh ảnh')}}>
+            <CardComponent style={{ alignItems: 'center' }} onPress={() => { Alert.alert('chỉnh ảnh') }}>
                 <View style={{ borderRadius: 10, overflow: 'hidden' }}>
                     <Image key={index}
                         source={{ uri: item?.photoUrl }}
@@ -217,29 +281,37 @@ const AddNewCustomerScreen = ({ navigation, route }) => {
 
     return (
         <ContainerComponent isScroll>
-            <HeaderComponent text={actionType === 'create' ? 'Thêm khách hàng mới' : 'Cập nhật thông tin khách hàng'} isBack
-                buttonRight={<Trash size={20} color={appColors.danger} />}
-                onRightPress={handleDeleteCustomer}
-            />
+            {actionType === 'create'
+                ? <HeaderComponent text='Thêm khách hàng' isBack />
+                :
+                <HeaderComponent text='Thông tin khách hàng' isBack
+                    buttonRight={<Trash size={20} color={appColors.danger} />}
+                    onRightPress={handleDeleteCustomer}
+                />}
 
             <SectionComponent style={{ alignItems: 'center' }}>
-                {/* <ImagePickerComponent text={actionType === 'create' ? 'Thêm ảnh minh hoạ' : 'Thay đổi ảnh minh hoạ'} onSelect={(val) => { setImageSelected(val) }} />
-          <SpaceComponent height={8} />
-          {imageSelected
-            ? <Image source={{ uri: imageSelected?.path }} style={{ height: 150, width: '100%', borderRadius: 10 }} resizeMode='cover' />
-            : dataRoom.photoUrl && <Image source={{ uri: dataRoom?.photoUrl }} style={{ height: 150, width: '100%', borderRadius: 10 }} resizeMode='cover' />
-          } */}
+                {/* ẩn này */}
+                <ImagePickerComponent ref={refImage} text={'Thêm ảnh minh hoạ'} onSelect={(val) => { setImageSelected(val) }} style={{ opacity: 0, position: 'absolute' }} />
                 <CircleComponent size={96} style={{ overflow: 'visible' }} >
-                    <Image source={{ uri: 'https://img.freepik.com/free-psd/3d-illustration-person-with-sunglasses_23-2149436178.jpg?size=338&ext=jpg&ga=GA1.1.2008272138.1726444800&semt=ais_hybrid' }}
-                        style={{ height: 96, width: 96, borderRadius: 999 }} resizeMode='cover' />
+                    {
+                        imageSelected
+                            ?
+                            <Image source={{ uri: imageSelected.path }} style={{ height: 96, width: 96, borderRadius: 999 }} resizeMode='cover' />
+                            : (dataCustomer.photoUrl
+                                ?
+                                <Image source={{ uri: dataCustomer.photoUrl }} style={{ height: 96, width: 96, borderRadius: 999 }} resizeMode='cover' />
+                                :
+                                <Image source={images.avatar_null}
+                                    style={{ height: 96, width: 96, borderRadius: 999 }} resizeMode='cover' />)
+                    }
                     <CircleComponent size={35} style={{ position: 'absolute', top: 0, left: appInfors.sizes.WIDTH * 0.17, backgroundColor: appColors.primary }}
-                        onPress={() => Alert.alert('edit image')}>
+                        onPress={() => { refImage.current.open() }}>
                         <Edit size={22} color={appColors.white} variant='Broken' />
                     </CircleComponent>
                 </CircleComponent>
 
                 <SpaceComponent height={10} />
-                <TextComponent text='Khánh Duy' isTitle />
+                {dataCustomer.customerName && <TextComponent text={dataCustomer.customerName} isTitle />}
 
             </SectionComponent>
 
@@ -286,18 +358,74 @@ const AddNewCustomerScreen = ({ navigation, route }) => {
 
                 <SpaceComponent height={14} />
 
-                <TextComponent text='Chọn ngày sinh here' />
+                <DateTimePickerComponent type='date' title='Chọn ngày sinh' selected={dataCustomer.dateOfBirth} onSelect={(val) => { handleChangeValue('dateOfBirth', val) }} />
             </SectionComponent>
 
             <SpaceComponent height={14} />
 
+            {/* vùng này của ảnh CCCD */}
             <SectionComponent>
-                <FlatList
+                {/* <FlatList
                     horizontal
                     data={testData}
                     renderItem={renderTestData}
                     showsHorizontalScrollIndicator={false}
-                />
+                /> */}
+                <ImagePickerComponent text={'Thêm ảnh CCCD mặt trước'} onSelect={(val) => { setImageSelected2(val) }}
+                    style={{ color: appColors.primary, textDecorationLine: 'underline' }} />
+                <SpaceComponent height={8} />
+                {
+                    imageSelected2
+                        ?
+                        <Image
+                            source={{ uri: imageSelected2.path }}
+                            style={{ height: 130, width: '100%' }}
+                            resizeMode='stretch'
+                        />
+                        :
+                        (dataCustomer.citizenIdphotoFirstUrl
+                            ?
+                            <Image
+                                source={{ uri: dataCustomer.citizenIdphotoFirstUrl }}
+                                style={{ height: 130, width: '100%' }}
+                                resizeMode='stretch'
+                            />
+                            :
+                            <Image
+                                source={images.cccd}
+                                style={{ height: 130, width: '100%' }}
+                                resizeMode='stretch'
+                            />
+                        )
+                }
+                <SpaceComponent height={8} />
+                <ImagePickerComponent text={'Thêm ảnh CCCD mặt sau'} onSelect={(val) => { setImageSelected3(val) }}
+                    style={{ color: appColors.primary, textDecorationLine: 'underline' }} />
+                <SpaceComponent height={8} />
+                {
+                    imageSelected3
+                        ?
+                        <Image
+                            source={{ uri: imageSelected3.path }}
+                            style={{ height: 130, width: '100%' }}
+                            resizeMode='stretch'
+                        />
+                        :
+                        (dataCustomer.citizenIdphotoBackUrl
+                            ?
+                            <Image
+                                source={{ uri: dataCustomer.citizenIdphotoBackUrl }}
+                                style={{ height: 130, width: '100%' }}
+                                resizeMode='stretch'
+                            />
+                            :
+                            <Image
+                                source={images.cccd}
+                                style={{ height: 130, width: '100%' }}
+                                resizeMode='stretch'
+                            />
+                        )
+                }
             </SectionComponent>
 
             <SpaceComponent height={14} />
@@ -311,6 +439,7 @@ const AddNewCustomerScreen = ({ navigation, route }) => {
                         <ButtonComponent text='Cập nhật khách hàng' onPress={handleUpdateCustomer} />
                 }
             </SectionComponent>
+            <SpaceComponent height={30} />
             <LoadingModalComponent visible={isLoading} />
         </ContainerComponent>
     )
